Add files via upload
Finished VigenereCipher.py
You can decrypt and encrypt what you want without redirection by just running the normal command and simply typing in whatever you want, as long as you provide a mode and key.
To bring in another file, it's necessary when sending an encrypted statement to type out what you want, or paste it into the terminal, and then the encrypted statement will be sent to the file.  For example: "python VigenereCipher.py -e "MyKey" > encrypt.txt" will send the encrypted statement to encrypt.txt, but you will need to type in whatever it is you want to encrypt.  For decryption, all you have to do is change the direction of the redirection symbol and change mode to decryption and the file will be input to the program as stdin and print out the decrypted message.